db wants to be a module global \o/
