+++ examples/todomvc/src/hooks/useUploader.tsx
@@ -1,4 +1,4 @@
-import { useEffect } from 'react'
+import { useEffect, useState } from 'react'
 import { useKeyring } from '@w3ui/react-keyring'
 import { store } from '@web3-storage/capabilities/store'
 import { Fireproof } from '@fireproof/core'
@@ -29,4 +29,63 @@ export function useUploader(database: Fireproof) {
     }
   }, [registered])
   return registered
+}
+
+export const UploadManager = ({ registered }: { registered: Boolean }) => {
+  if (registered) {
+    return <p>Your changes are being saved to the public IPFS network with web3.storage</p>
+  } else {
+    return <SpaceRegistrar />
+  }
+}
+
+function SpaceRegistrar(): JSX.Element {
+  const [, { registerSpace }] = useKeyring()
+  const [email, setEmail] = useState('')
+  const [submitted, setSubmitted] = useState(false)
+  function resetForm(): void {
+    setEmail('')
+  }
+  async function onSubmit(e: React.FormEvent<HTMLFormElement>): Promise<void> {
+    e.preventDefault()
+    setSubmitted(true)
+    try {
+      await registerSpace(email)
+    } catch (err) {
+      console.log(err)
+      throw new Error('failed to register', { cause: err })
+    } finally {
+      resetForm()
+      setSubmitted(false)
+    }
+  }
+  return (
+    <div className="flex flex-col items-center space-y-24 pt-12">
+      <div className="flex flex-col items-center space-y-2">
+        <h3 className="text-lg">Verify your email address!</h3>
+        <p>web3.storage is sending you a verification email. Please click the link.</p>
+      </div>
+      <div className="flex flex-col items-center space-y-4">
+        <h5>Need a new verification email?</h5>
+        <form
+          className="flex flex-col items-center space-y-2"
+          onSubmit={(e: React.FormEvent<HTMLFormElement>) => {
+            void onSubmit(e)
+          }}
+        >
+          <input
+            className="text-black px-2 py-1 rounded"
+            type="email"
+            placeholder="Email"
+            value={email}
+            onChange={(e) => {
+              setEmail(e.target.value)
+            }}
+          />
+          <input type="submit" className="w3ui-button" value="Re-send Verification Email" disabled={email === ''} />
+        </form>
+        {submitted && <p>Verification re-sent, please check your email for a verification email.</p>}
+      </div>
+    </div>
+  )
 }
